Made tokenAuth as a middleware, updated userControl and userRoute, successfully handled token auth
